For merge

+++ challenge/views.py
@@ -1,13 +1,7 @@
-from django.shortcuts import render, redirect
+from django.shortcuts import render, redirect, get_object_or_404
 from django.http import HttpResponse
-from .models import ChallengeModel
+from .models import ChallengeModel, ChallengeJoinModel
 from django.contrib.auth.decorators import login_required
-
-
-def delete_challenge(request, id):
-    target_challenge = ChallengeModel.objects.get(id=id)
-    target_challenge.delete()
-    return redirect('/main')
 
 
 def view_main(request):
@@ -42,16 +36,25 @@ def posting_challenge(request):
             return redirect('/challenge/' + str(challenge_id))
         else:
             pass
+    else:
+        return redirect('/main')
 
 
 @login_required
 def challenge_detail(request, id):
-    target_challenge = ChallengeModel.objects.get(id=id)
-    if request.method == 'GET':
-        context = {
-            'challenge': target_challenge
-        }
-        return render(request, 'challenge/detail.html', context)
+    target_challenge = get_object_or_404(ChallengeModel, id=id)
+    user = request.user
+    is_joined = ChallengeJoinModel.objects.filter(
+        joined_challenge=target_challenge, joined_user=user).exists()
+    is_completed = ChallengeJoinModel.objects.filter(
+        joined_challenge=target_challenge, joined_user=user, complete=True).exists()
+
+    context = {
+        'challenge': target_challenge,
+        'is_joined': is_joined,
+        'is_completed': is_completed,
+    }
+    return render(request, 'challenge/detail.html', context)
 
 
 # =========챌린지 검색 view ============
@@ -64,28 +67,72 @@ def challenge_search_view(request):
 
 @login_required
 def delete_challenge(request, id):
-    target_challenge = ChallengeModel.objects.get(id=id)
-    target_challenge.delete()
+    target_challenge = get_object_or_404(ChallengeModel, id=id)
+    if request.user == target_challenge.challenge_author:
+        target_challenge.delete()
+    else:
+        return HttpResponse("권한이 없습니다.")
     return redirect('/main')
 
 
 @login_required
 def edit_challenge(request, id):
-    target_challenge = ChallengeModel.objects.get(id=id)
-    if request.method == 'GET':
-        context = {
-            'challenge': target_challenge
-        }
-        return render(request, 'challenge/detail_edit.html', context)
+    target_challenge = get_object_or_404(ChallengeModel, id=id)
+    if request.user == target_challenge.challenge_author:
+        if request.method == 'GET':
+            context = {
+                'challenge': target_challenge
+            }
+            return render(request, 'challenge/detail_edit.html', context)
 
-    elif request.method == 'POST':
-        target_challenge.challenge_title = request.POST.get('challenge_title')
-        target_challenge.challenge_name = request.POST.get('challenge_name')
-        target_challenge.challenge_genre = request.POST.get('challenge_genre')
-        target_challenge.challenge_content = request.POST.get(
-            'challenge_content')
-        target_challenge.challenge_image = request.FILES.get('challenge_image')
-        if target_challenge.challenge_title and target_challenge.challenge_name and target_challenge.challenge_genre and target_challenge.challenge_content:
-            target_challenge.save()
-            challenge_id = target_challenge.id
-            return redirect('/challenge/' + str(challenge_id))
+        elif request.method == 'POST':
+            target_challenge.challenge_title = request.POST.get(
+                'challenge_title')
+            target_challenge.challenge_name = request.POST.get(
+                'challenge_name')
+            target_challenge.challenge_genre = request.POST.get(
+                'challenge_genre')
+            target_challenge.challenge_content = request.POST.get(
+                'challenge_content')
+            target_challenge.challenge_image = request.FILES.get(
+                'challenge_image')
+            if target_challenge.challenge_title and target_challenge.challenge_name and target_challenge.challenge_genre and target_challenge.challenge_content:
+                target_challenge.save()
+                challenge_id = target_challenge.id
+                return redirect('/challenge/' + str(challenge_id))
+    else:
+        return HttpResponse("권한이 없습니다.")
+
+
+def join_challenge(request, id):
+    target_challenge = get_object_or_404(ChallengeModel, id=id)
+    target_user = request.user
+
+    # 챌린지에 이미 참가한 경우
+    if ChallengeJoinModel.objects.filter(joined_challenge=target_challenge, joined_user=target_user).exists():
+        return HttpResponse("이미 참가중입니다.")
+
+    # 챌린지에 아직 참가하지 않은 경우
+    ChallengeJoinModel.objects.create(
+        joined_challenge=target_challenge, joined_user=target_user)
+
+    return HttpResponse("참가 완료")
+
+
+def complete_challenge(request, id):
+    target_challenge = get_object_or_404(ChallengeModel, id=id)
+    target_user = request.user
+
+    try:
+        joined_challenge = ChallengeJoinModel.objects.get(
+            joined_challenge=target_challenge, joined_user=target_user)
+    except ChallengeJoinModel.DoesNotExist:
+        return HttpResponse("참가부터 하세요.")
+
+    # 참가한 상태에서만 완료하게 하는 판별식
+    if not joined_challenge.complete:
+        joined_challenge.complete = True
+        joined_challenge.save()
+        return HttpResponse("챌린지 완료")
+    else:
+        return HttpResponse("이미 완료한 챌린지입니다.")
